% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{distance_weights}
\alias{distance_weights}
\title{Distance weight methods}
\usage{
distance_weights(D, d_max, imp_function, function_d_max = 0.01)
}
\arguments{
\item{D}{numeric matrix, distance or time values}

\item{d_max}{numeric, threshold for max distance}

\item{imp_function}{character, type of distance weights method}

\item{function_d_max}{numeric, result for function(d_max)}
}
\value{
matrix, normalized distance or time values
}
\description{
Distance weight methods
}
\examples{
distance_weights(matrix(10), 10, "gaussian")
}
