% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_2SFCA.R
\name{get_2SFCA}
\alias{get_2SFCA}
\title{Two-Step Floating Catchment Area method}
\usage{
get_2SFCA(p, s, D, step)
}
\arguments{
\item{p}{numeric vector, number of population at origin locations}

\item{s}{numeric vector, capacity of services at supply locations}

\item{D}{matrix, distance or time matrix}

\item{step}{character, step of the method to perform}
}
\value{
data.frame, depending on selected step
}
\description{
Two-Step Floating Catchment Area method
}
\examples{
p <- 1:4
s <- 1:6
D <- matrix(1:24, ncol=4, nrow=6)
spai <- get_2SFCA(p, s, D, "step2")
}
