% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{get_distance_weights}
\alias{get_distance_weights}
\title{Distance weights methods}
\usage{
get_distance_weights(D, d0, impFunction, functiond0 = 0.01)
}
\arguments{
\item{D}{matrix, distance or time values}

\item{d0}{numeric, threshold for max distance}

\item{impFunction}{character, type of distance weights method}

\item{functiond0}{numeric, result for function(d0)}
}
\value{
matrix, normalized distance or time values
}
\description{
Distance weights methods
}
\examples{
get_distance_weights(matrix(10), 10, "gaussian")
}
